+++ router/core/router.go
@@ -10,6 +10,11 @@ import (
 	"sync"
 	"time"
 
+	"connectrpc.com/connect"
+	"github.com/wundergraph/cosmo/router/gen/proto/wg/cosmo/graphqlmetrics/v1/graphqlmetricsv1connect"
+	"github.com/wundergraph/cosmo/router/internal/graphqlmetrics"
+	brotli "go.withmatt.com/connect-brotli"
+
 	"github.com/wundergraph/cosmo/router/internal/otel/otelconfig"
 
 	"github.com/dgraph-io/ristretto"
@@ -59,6 +64,11 @@ type (
 		KeepAliveProbeInterval time.Duration
 	}
 
+	GraphQLMetricsConfig struct {
+		Enabled           bool
+		CollectorEndpoint string
+	}
+
 	// Config defines the configuration options for the Router.
 	Config struct {
 		transport                *http.Transport
@@ -67,6 +77,7 @@ type (
 		metricConfig             *metric.Config
 		tracerProvider           *sdktrace.TracerProvider
 		meterProvider            *sdkmetric.MeterProvider
+		gqlMetricsExporter       *graphqlmetrics.Exporter
 		corsOptions              *cors.Config
 		configFetcher            controlplane.ConfigFetcher
 		routerConfig             *nodev1.RouterConfig
@@ -77,7 +88,6 @@ type (
 		graphqlPath              string
 		playground               bool
 		introspection            bool
-		production               bool
 		federatedGraphName       string
 		graphApiToken            string
 		healthCheckPath          string
@@ -91,9 +101,9 @@ type (
 		headerRuleEngine         *HeaderRuleEngine
 		headerRules              config.HeaderRules
 		subgraphTransportOptions *SubgraphTransportOptions
+		graphqlMetricsConfig     *GraphQLMetricsConfig
 		routerTrafficConfig      *config.RouterTrafficConfiguration
-
-		retryOptions retrytransport.RetryOptions
+		retryOptions             retrytransport.RetryOptions
 
 		engineExecutionConfiguration config.EngineExecutionConfiguration
 
@@ -152,6 +162,9 @@ func NewRouter(opts ...Option) (*Router, error) {
 		r.subgraphTransportOptions = DefaultSubgraphTransportOptions()
 	}
 
+	if r.graphqlMetricsConfig == nil {
+		r.graphqlMetricsConfig = DefaultGraphQLMetricsConfig()
+	}
 	if r.routerTrafficConfig == nil {
 		r.routerTrafficConfig = DefaultRouterTrafficConfig()
 	}
@@ -276,13 +289,7 @@ func (r *Router) configureSubgraphOverwrites(cfg *nodev1.RouterConfig) ([]Subgra
 
 			// Override datasource urls
 			for _, conf := range cfg.EngineConfig.DatasourceConfigurations {
-				fetchURL := conf.CustomGraphql.Fetch.Url
-				subgraphURL := config.LoadStringVariable(fetchURL)
-
-				// Identify the datasource by the previous subgraph url
-				// Override datasource id, url and subgraph url
-				if subgraphURL == sg.RoutingUrl {
-					conf.Id = overrideURL
+				if conf.Id == sg.Id {
 					conf.CustomGraphql.Fetch.Url.StaticVariableContent = overrideURL
 					conf.CustomGraphql.Subscription.Url.StaticVariableContent = overrideURL
 					sg.RoutingUrl = overrideURL
@@ -455,6 +462,29 @@ func (r *Router) bootstrap(ctx context.Context) error {
 		}
 	}
 
+	if r.graphqlMetricsConfig.Enabled {
+		client := graphqlmetricsv1connect.NewGraphQLMetricsServiceClient(
+			http.DefaultClient,
+			r.graphqlMetricsConfig.CollectorEndpoint,
+			brotli.WithCompression(),
+			// Compress requests with Brotli.
+			connect.WithSendCompression(brotli.Name),
+		)
+		r.gqlMetricsExporter = graphqlmetrics.NewExporter(
+			r.logger,
+			client,
+			r.graphApiToken,
+			graphqlmetrics.NewDefaultExporterSettings(),
+		)
+		if err := r.gqlMetricsExporter.Validate(); err != nil {
+			return fmt.Errorf("failed to validate graphql metrics exporter: %w", err)
+		}
+
+		r.gqlMetricsExporter.Start()
+
+		r.logger.Info("GraphQL schema coverage metrics enabled")
+	}
+
 	// Modules are only initialized once and not on every config change
 	if err := r.initModules(ctx); err != nil {
 		return fmt.Errorf("failed to init user modules: %w", err)
@@ -587,6 +617,7 @@ func (r *Router) newServer(ctx context.Context, routerConfig *nodev1.RouterConfi
 		baseURL:       r.baseURL,
 		transport:     r.transport,
 		logger:        r.logger,
+		includeInfo:   r.graphqlMetricsConfig.Enabled,
 		transportOptions: &TransportOptions{
 			requestTimeout: r.subgraphTransportOptions.RequestTimeout,
 			preHandlers:    r.preOriginHandlers,
@@ -609,7 +640,8 @@ func (r *Router) newServer(ctx context.Context, routerConfig *nodev1.RouterConfi
 		return nil, fmt.Errorf("failed to build plan configuration: %w", err)
 	}
 
-	operationParser := NewOperationParser(executor)
+	operationParser := NewOperationParser(executor, int64(r.routerTrafficConfig.MaxRequestBodyBytes))
+	operationPlanner := NewOperationPlanner(executor, planCache)
 
 	var graphqlPlaygroundHandler http.Handler
 
@@ -624,7 +656,6 @@ func (r *Router) newServer(ctx context.Context, routerConfig *nodev1.RouterConfi
 
 	graphqlHandler := NewGraphQLHandler(HandlerOptions{
 		Executor: executor,
-		Cache:    planCache,
 		Log:      r.logger,
 	})
 
@@ -648,11 +679,14 @@ func (r *Router) newServer(ctx context.Context, routerConfig *nodev1.RouterConfi
 		metricStore = m
 	}
 
+	routerMetrics := NewRouterMetrics(metricStore, r.gqlMetricsExporter, routerConfig.GetVersion())
+
 	graphqlPreHandler := NewPreHandler(&PreHandlerOptions{
-		Parser:                operationParser,
-		Logger:                r.logger,
-		RequestMetrics:        metricStore,
-		MaxRequestSizeInBytes: int64(r.routerTrafficConfig.MaxRequestBodyBytes),
+		Logger:   r.logger,
+		Executor: executor,
+		Metrics:  routerMetrics,
+		Parser:   operationParser,
+		Planner:  operationPlanner,
 	})
 
 	var traceHandler *trace.Middleware
@@ -681,10 +715,11 @@ func (r *Router) newServer(ctx context.Context, routerConfig *nodev1.RouterConfi
 		}
 
 		subChiRouter.Use(NewWebsocketMiddleware(rootContext, WebsocketMiddlewareOptions{
-			Parser:                operationParser,
-			MaxRequestSizeInBytes: int64(r.routerTrafficConfig.MaxRequestBodyBytes),
-			GraphQLHandler:        graphqlHandler,
-			Logger:                r.logger,
+			Parser:         operationParser,
+			Planner:        operationPlanner,
+			Metrics:        routerMetrics,
+			GraphQLHandler: graphqlHandler,
+			Logger:         r.logger,
 		}))
 
 		subChiRouter.Use(graphqlPreHandler.Handler)
@@ -735,6 +770,12 @@ func (r *Server) listenAndServe() error {
 func (r *Router) Shutdown(ctx context.Context) (err error) {
 	r.shutdown = true
 
+	if r.activeRouter != nil {
+		if subErr := r.activeRouter.Shutdown(ctx); subErr != nil {
+			err = errors.Join(err, fmt.Errorf("failed to shutdown primary server: %w", subErr))
+		}
+	}
+
 	var wg sync.WaitGroup
 
 	if r.prometheusServer != nil {
@@ -753,11 +794,20 @@ func (r *Router) Shutdown(ctx context.Context) (err error) {
 		go func() {
 			defer wg.Done()
 
-			if subErr := r.tracerProvider.ForceFlush(ctx); subErr != nil {
-				err = errors.Join(err, fmt.Errorf("failed to force flush tracer: %w", subErr))
-			}
 			if subErr := r.tracerProvider.Shutdown(ctx); subErr != nil {
 				err = errors.Join(err, fmt.Errorf("failed to shutdown tracer: %w", subErr))
+			}
+		}()
+	}
+
+	if r.gqlMetricsExporter != nil {
+		wg.Add(1)
+
+		go func() {
+			defer wg.Done()
+
+			if subErr := r.gqlMetricsExporter.Shutdown(ctx); subErr != nil {
+				err = errors.Join(err, fmt.Errorf("failed to stop graphql metrics exporter: %w", subErr))
 			}
 		}()
 	}
@@ -776,12 +826,6 @@ func (r *Router) Shutdown(ctx context.Context) (err error) {
 	}()
 
 	wg.Wait()
-
-	if r.activeRouter != nil {
-		if subErr := r.activeRouter.Shutdown(ctx); subErr != nil {
-			err = errors.Join(err, fmt.Errorf("failed to shutdown primary server: %w", subErr))
-		}
-	}
 
 	return err
 }
@@ -826,7 +870,7 @@ func createPrometheus(logger *zap.Logger, listenAddr, path string) *http.Server 
 		Handler:           r,
 	}
 
-	logger.Info("Serve Prometheus metrics", zap.String("listen_addr", svr.Addr), zap.String("endpoint", path))
+	logger.Info("Prometheus metrics enabled", zap.String("listen_addr", svr.Addr), zap.String("endpoint", path))
 
 	return svr
 }
@@ -1003,5 +1047,18 @@ func DefaultSubgraphTransportOptions() *SubgraphTransportOptions {
 		KeepAliveProbeInterval: 30 * time.Second,
 		KeepAliveIdleTimeout:   0 * time.Second,
 		DialTimeout:            30 * time.Second,
+	}
+}
+
+func DefaultGraphQLMetricsConfig() *GraphQLMetricsConfig {
+	return &GraphQLMetricsConfig{
+		Enabled:           false,
+		CollectorEndpoint: "",
+	}
+}
+
+func WithGraphQLMetrics(cfg *GraphQLMetricsConfig) Option {
+	return func(r *Router) {
+		r.graphqlMetricsConfig = cfg
 	}
 }