+++ studio/CHANGELOG.md
@@ -4,6 +4,16 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.27.0](https://github.com/wundergraph/cosmo/compare/studio@0.26.0...studio@0.27.0) (2023-10-25)
+
+### Bug Fixes
+
+* always show top 5 in metrics overview ([#203](https://github.com/wundergraph/cosmo/issues/203)) ([ca4f67d](https://github.com/wundergraph/cosmo/commit/ca4f67d93f192028ce4025893ac5d7ea18f299b7)) (@Pagebakers)
+
+### Features
+
+* schema field level usage analytics ([#174](https://github.com/wundergraph/cosmo/issues/174)) ([4f257a7](https://github.com/wundergraph/cosmo/commit/4f257a71984e991be2304b09a083c69da65200d2)) (@StarpTech)
+
 # [0.26.0](https://github.com/wundergraph/cosmo/compare/studio@0.25.0...studio@0.26.0) (2023-10-23)
 
 ### Features